Bugfix: Moved prefix filtering to theorem pruning.
If pruning happens in the prune_bounds code, the result is considered to be incomplete.
However, the idea of prefix filtering is that the user is aware that he is only considering
part of the search space, and we want to let him know if this part was fully explored or not.
Thus, we move prefix filtering to prume_theorems, essentially making it an environmental
assumption on the search space and therefore a premise of the proof (if any).